1.3.0 release candidate RC3

Conflicts:
	resource/c_common/octhread/src/posix/octhread.c
	resource/c_common/octhread/src/windows/octhread.c
	resource/csdk/security/src/credresource.c
	run.bat

+++ resource/csdk/routing/include/routingtablemanager.h
@@ -287,7 +287,7 @@ void RTMFreeGateway(RTMGatewayId_t *gateway, u_linklist_t **gatewayTable);
  * @param[in,out]    obsListLen             Length if Observation ID list.
  * @param[in]        gatewayTable           Gateway Routing Table.
  */
-void RTMGetObserverList(OCObservationId **obsList, uint8_t *obsListLen,
+void RTMGetObserverList(OCObservationId **obsList, uint32_t *obsListLen,
                         const u_linklist_t *gatewayTable);
 
 /**
